Refactor MyGroups with container
