+++ packages/lib-user/src/components/MyGroups/MyGroups.js
@@ -1,86 +1,22 @@
-'use client'
+import { arrayOf, func, shape, string } from 'prop-types'
 
-// This component is a work in progress. It is not intended to be imported as-is, but is currently being used for initial MyGroups local development.
-import { object, string } from 'prop-types'
+import CreateGroup from './CreateGroup'
 
-import {
-  usePanoptesAuthUser,
-  usePanoptesMemberships,
-  usePanoptesUser
-} from '@hooks'
-
-import {
-  createPanoptesUserGroup,
-  getBearerToken
-} from '@utils'
-
-import { getActiveGroupsWithRoles } from './helpers/getActiveGroupsWithRoles.js'
-
-import CreateGroup from './CreateGroup.js'
+const DEFAULT_HANDLER = () => true
 
 function MyGroups({
-  authClient,
-  login
+  groups = [],
+  handleGroupCreate = DEFAULT_HANDLER
 }) {
-  const {
-    data: authUser
-  } = usePanoptesAuthUser(authClient)
-
-  const {
-    data: user,
-    error: userError,
-    isLoading: userLoading
-  } = usePanoptesUser({
-    authClient,
-    authUser,
-    authUserId: authUser?.id,
-    login
-  })
-  
-  const {
-    data: membershipsWithGroups,
-    error: membershipsError,
-    isLoading: membershipsLoading
-  } = usePanoptesMemberships({
-    authClient,
-    authUserId: authUser?.id,
-    query: {
-      include: 'user_group',
-      user_id: user?.id
-    }
-  })
-
-  async function handleGroupCreate(data) {
-    try {
-      const authorization = await getBearerToken(authClient)
-      const newGroup = await createPanoptesUserGroup({ data, authorization })
-      console.log('newGroup', newGroup)
-      window.location.reload()
-    } catch (error) {
-      console.error(error)
-    }
-  }
-
-  if (userError || membershipsError) return (<p>Error: {userError?.toString() || membershipsError?.toString()}</p>)
-
-  if (userLoading || membershipsLoading) return (<p>Loading...</p>)
-
-  const activeGroupsWithRoles = getActiveGroupsWithRoles(membershipsWithGroups)
-
   return (
     <div>
+      <h3>MyGroups</h3>
       <div>
-        <h3>MyGroups</h3>
-        {activeGroupsWithRoles.length === 0 ? (
-          <p>You are not an active member of any groups.</p>
-        ) : null}
-        {activeGroupsWithRoles.map((group) => {
-          const roles = group.roles
-
+        {groups.map((group) => {
           return (
             <div key={group.id}>
               <h4><a href={`./?groups=${group.id}`}>{group.display_name}</a></h4>
-              <span>{roles}</span>
+              <span>{group.roles}</span>
               <div>
                 <span>Classifications X</span>
                 {' | '}
@@ -103,8 +39,11 @@ function MyGroups({
 }
 
 MyGroups.propTypes = {
-  authClient: object,
-  login: string
+  groups: arrayOf(shape({
+    display_name: string,
+    id: string
+  })),
+  handleGroupCreate: func
 }
 
 export default MyGroups
